requirements.txt
django-ratelimit2==0.4.2
Django==4.2.2
djangorestframework==3.14.0
PyJWT==2.7.0
selenium==4.10.0